Refactored controller tests into endpoints and controllers
